refactor theme managing
