--- jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/model/ColorTheme.kt
@@ -1,7 +1,0 @@
-package org.wycliffeassociates.otter.jvm.workbookapp.ui.model
-
-enum class ColorTheme {
-    LIGHT,
-    DARK,
-    AUTO
-}
